+++ src/main/java/org/example/riotspring/riot/lol/entries/EntryController.java
@@ -1,15 +1,16 @@
-package org.example.riotspring.entries;
+package org.example.riotspring.riot.lol.entries;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.example.riotspring.RiotAPIService;
-import org.example.riotspring.RiotApiDTO;
+import org.example.riotspring.riot.RiotAPIService;
+import org.example.riotspring.riot.RiotApiDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.example.riotspring.riot.lol.summoner.entries.LeagueEntryDTO;
 
 @RestController
 @RequestMapping("/api/entries")
